boxplot2 {gplots}	R Documentation

Produce a Boxplot Annotated with the Number of Observations

Description

This funcntion uses boxplot to produce a boxplot which is then annotated with the number of observations in each group.

Usage

boxplot2(..., top=FALSE, shrink=1, textcolor=NULL)

Arguments

...
parameters passed to boxplot.

top
logical indicating whether the number of observations should be added to the top or the bottom of the plotting region. Defaults to FALSE.

shrink
value to shrink character size (cex) when annotating.

textcolor
text color.

Note

This function replaces boxplot.n, which has been deprecated avoid potential problems with S3 method dispatching.

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

boxplot, text

Examples

data(state)

# n's at bottom
boxplot2( state.area ~ state.region)

# n's at top
boxplot2( state.area ~ state.region, top=TRUE)

# small red text
boxplot2( state.area ~ state.region, shrink=0.8, textcolor="red") 

[Package gplots version 3.1.3 Index]