catch.d {gplots}	R Documentation

Catch at Age and Residuals

Description

Catch-at-age observed data and model residuals from Icelandic saithe assessment.

Usage

catch.d
catch.r

Format

Data frame containing three columns:

Year 	 year

Age 	 age

and

Catch 	 catch (thousands of individuals)

or

Resid 	 standardized residual

Details

The data are from Tables 8.2 and 8.6 in the ICES (2015) fish stock assessment of Icelandic saithe.

Source

ICES (2015) Report of the North-Western Working Group (NWWG). ICES CM 2015/ACOM:07, pp. 240–246.

See Also

bubbleplot is an effective way to visualize these data.

Examples

catch.t <- xtabs(Catch~Year+Age, catch.d)
catch.m <- as.matrix(as.data.frame(unclass(catch.t)))

# 1  Formula
bubbleplot(Catch~Age+Year, data=catch.d)
# Use rev=TRUE to get same layout as crosstab matrix:
print(catch.m)
bubbleplot(Catch~Age+Year, data=catch.d, rev=TRUE, las=1)

# 2  Data frame
bubbleplot(catch.d)

# 3  Matrix or table
bubbleplot(catch.m)
bubbleplot(catch.t)

# 4  Positive and negative values
bubbleplot(catch.r)
bubbleplot(Resid~Age+Year, catch.r, subset=Age %in% 4:9,
           rev=TRUE, xlim=c(3.5,9.5), cex=1.3)

[Package gplots version 3.3.0 Index]