gplots-deprecated {gplots}	R Documentation

Deprecated functions

Description

These functions have been deprecated and will be removed in future releases of gplots.

Usage

  ## No deprecated functions at this time ##

Details

These functions have been deprecated. Please refer to the manual page for the replacement function:

(No deprecated functions at this time)

Author(s)

Gregory R. Warnes greg@warnes.net

See Also

Deprecated

[Package gplots version 3.1.3 Index]