residplot {gplots}	R Documentation

Undocumented functions

Description

These functions are undocumented. Some are internal and not intended for direct use. Others simply haven't been documented yet.

Usage

residplot(model, formula, ...) 

Arguments

model
Undocumented

formula
Undocumented

...
arguments to be passed to fun

Details

These functions are undocumented. Some are internal and not intended for direct use. Others simply haven't been documented yet.

Author(s)

Gregory R. Warnes greg@warnes.net

[Package gplots version 3.1.3 Index]