Defunct functions

These functions are defunct and have been removed from the gplots package.

boxplot.n(..., top=FALSE, shrink=1, textcolor=NULL)
   plot.lm2(
            x,
            which = 1:5,
            caption = c("Residuals vs Fitted", "Normal Q-Q plot",
              "Scale-Location plot", "Cook's distance plot"),
            panel = panel.smooth,
            sub.caption = deparse(x$call),
            main = "",
            ask,
            ...,
            id.n = 3,
            labels.id = names(residuals(x)),
            cex.id = 0.75,
            band=TRUE,
            rug=TRUE,
            width=1/10,
            max.n=5000
            )
  smartlegend(x = c("left", "center", "right"),
             y = c("top", "center", "bottom"),
             ...,
             inset = 0.05)

Arguments

ask, band, caption, cex.id, id.n, inset, labels.id, main, max.n, panel, rug, shrink, sub.caption, textcolor, top, which, width, x, y, ...
see man page for the corresponding replacement function

Details

These functions are no longer available. Please refer to the manual page for the replacement function:

boxplot.n has been replaced by boxplot2

plot.lm2 has been replaced by lmplot2

smartlegend is no longer needed because relative positioning has been implemented in legend.

See also