Scatter Plot Smoothing

The lowess function performs the computations for the LOWESS smoother (see the reference below). lowess returns a an object containing components x and y which give the coordinates of the smooth. The smooth can then be added to a plot of the original points with the function lines.

lowess(x, ...)

Arguments

x, y

vectors giving the coordinates of the points in the scatter plot. Alternatively a single plotting structure can be specified.

...

parameters for methods.

formula

formula providing a single dependent variable (y) and an single independent variable (x) to use as coordinates in the scatter plot.

data

a data.frame (or list) from which the variables in `formula' should be taken.

subset

an optional vector specifying a subset of observations to be used in the fitting process.

f

the smoother span. This gives the proportion of points in the plot which influence the smooth at each value. Larger values give more smoothness.

iter

the number of robustifying iterations which should be performed. Using smaller values of iter will make lowess run faster.

delta

values of x which lie within delta of each other replaced by a single value in the output from lowess.

col.lowess, lty.lowess

color and line type for plotted line

Details

Alternatively, plot can be called directly on the object returned from lowess and the 'lowess' method for plot will generate a scatterplot of the original data with a lowess line superimposed.

Finally, the plotLowess function both calculates the lowess smooth and plots the original data with a lowess smooth.

References

Cleveland, W. S. (1979) Robust locally weighted regression and smoothing scatterplots. J. Amer. Statist. Assoc. 74, 829–836.

Cleveland, W. S. (1981) LOWESS: A program for smoothing scatterplots by robust locally weighted regression. The American Statistician, 35, 54.

See also

Examples