sinkplot {gplots}	R Documentation

Send textual R output to a graphics device

Description

Divert R's standard text output to a graphics device.

Usage

sinkplot(operation = c("start", "plot", "cancel"), ...)

Arguments

operation 	 See below

... 	 Plot arguments. (Ignored unless operation="plot").

Details

This function allows the printed output of R commands to be captured and displayed on a graphics device.

The capture process is started by calling sinkplot("start"). Now R commands can be executed and all printed output (except errors) will be captured. When the desired text has been captured sinkplot("plot") can be called to actually display the output. sinkplot("cancel") can be used to abort the output capture without plotting.

The current implementation does not allow sinkplot to be nested.

Value

Invisibly returns a character vector containing one element for each line of the captured output.

Author(s)

Gregory R. Warnes greg@random-technologies-llc.com

References

Functionality requested by Kevin Wright kwright@eskimo.com in the R-devel newlist posting https://www.stat.math.ethz.ch/pipermail/r-devel/2004-January/028483.html

See Also

capture.output, textplot

Examples

## Not run: 
   set.seed(12456)
   x <- factor(sample( LETTERS[1:5], 50, replace=TRUE))
   y <- rnorm(50, mean=as.numeric(x), sd=1)

   par(mfrow=c(1,2))
   boxplot(y~x, col="darkgreen")

   sinkplot()
   anova(lm(y~x))
   sinkplot("plot",col="darkgreen")
## End(Not run)

[Package gplots version 2.7.1 Index]